+++ src/love/image.cpp
@@ -1,5 +1,7 @@
 #include "image.h"
 #include <string>
+#include <list>
+#include <iostream>
 #include "Types/Graphics/Image.h"
 
 using love::Types::Graphics::Image;
@@ -9,8 +11,10 @@ namespace love {
 Image* image::newImageData(const std::string& filename) {
 	Image* image = new Image(filename);
 	if (image->loaded()) {
+		m_images.push_back(image);
 		return image;
 	}
+	delete image;
 	return NULL;
 }
 
@@ -33,7 +37,12 @@ bool image::load() {
 }
 
 bool image::unload() {
-	// IMG_Quit();
+	for (std::list<Image*>::iterator it = m_images.begin(); it != m_images.end(); ++it) {
+		if (*it != NULL) {
+	    	delete *it;
+	    }
+	}
+	m_images.clear();
 	return true;
 }