+++ src/love/system.cpp
@@ -2,8 +2,10 @@
 #include "../ChaiLove.h"
 
 #include <string>
-#include <cstring>
+#include <cstdlib>
 #include <iostream>
+#include <cstdio>
+#include <vector>
 #include "semver.h"
 #include "libretro.h"
 
@@ -61,6 +63,29 @@ std::string system::getVersionString() {
 }
 
 bool system::load(config& t) {
+	// Update core option from the libretro variables.
+	struct retro_variable var = {0};
+
+	// Alpha Blending
+	var.key = "chailove_alphablending";
+	var.value = NULL;
+	if (ChaiLove::environ_cb(RETRO_ENVIRONMENT_GET_VARIABLE, &var) && var.value) {
+		std::string varvalue(var.value);
+		if (varvalue == "disabled") {
+			t.options["alphablending"] = false;
+		}
+	}
+
+	// High Quality
+	var.key = "chailove_highquality";
+	var.value = NULL;
+	if (ChaiLove::environ_cb(RETRO_ENVIRONMENT_GET_VARIABLE, &var) && var.value) {
+		std::string varvalue(var.value);
+		if (varvalue == "disabled") {
+			t.options["highquality"] = false;
+		}
+	}
+
 	// Load the semantic version string.
 	semver_t chailoveVersion = {};
 	semver_t coreVersion = {};
@@ -120,6 +145,17 @@ std::string system::getUsername() {
 		}
 	}
 	return m_username;
+}
+
+bool system::execute(const std::string& command) {
+	std::cout << "[ChaiLove] [system] love.system.execute(\"" << command << "\")" << std::endl;
+	int result = std::system(command.c_str());
+	if (result != 0) {
+		std::cout << "[ChaiLove] [system] Failed to execute " << command << std::endl;
+		return false;
+	}
+	std::cout << "[ChaiLove] [system] Finished " << command << std::endl;
+	return true;
 }
 
 }  // namespace love